drop more unused database functions
